server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
redis
configparser
python-dotenv
spotipy